# server/portal/internal/docs/swagger.yaml
basePath: /fe-v1
definitions:
...
  render.ErrorResponse:
    properties:
      error:
        example: 操作失败
        type: string
    type: object
  render.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  service.CreateResourcePoolDeviceMatchingPolicyRequest:
    properties:
      actionType:
        description: 动作类型：pool_entry 或 pool_exit
        enum:
        - pool_entry
        - pool_exit
        type: string
      additionConds:
        description: 额外动态条件，仅入池时有效
        items:
          type: string
        type: array
      description:
        description: 策略描述
        type: string
      name:
        description: 策略名称
        type: string
      queryTemplateId:
        description: 关联的查询模板ID
        type: integer
      resourcePoolType:
        description: 资源池类型
        type: string
      status:
        description: 状态：enabled 或 disabled
        enum:
        - enabled
        - disabled
        type: string
    required:
    - actionType
    - name
    - queryTemplateId
    - resourcePoolType
    - status
    type: object
...
  service.UpdateResourcePoolDeviceMatchingPolicyRequest:
    properties:
      actionType:
        description: 动作类型：pool_entry 或 pool_exit
        enum:
        - pool_entry
        - pool_exit
        type: string
      additionConds:
        description: 额外动态条件，仅入池时有效
        items:
          type: string
        type: array
      description:
        description: 策略描述
        type: string
      name:
        description: 策略名称
        type: string
      queryTemplateId:
        description: 关联的查询模板ID
        type: integer
      resourcePoolType:
        description: 资源池类型
        type: string
      status:
        description: 状态：enabled 或 disabled
        enum:
        - enabled
        - disabled
        type: string
    required:
    - actionType
    - name
    - queryTemplateId
    - resourcePoolType
    - status
    type: object
  service.UpdateResourcePoolDeviceMatchingPolicyStatusRequest:
    properties:
      status:
        description: 状态：enabled 或 disabled
        enum:
        - enabled
        - disabled
        type: string
    required:
    - status
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Navy-NG 管理平台 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Navy-NG API
  version: "1.0"
paths:
...
  /fe-v1/devices:
    get:
      consumes:
      - application/json
      description: 获取设备列表，支持分页
      parameters:
      - description: 页码，默认1
        in: query
        name: page
        type: integer
      - description: 每页大小，默认10
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.ErrorResponse'
      summary: 获取设备列表
      tags:
      - 设备管理
  /fe-v1/devices/{id}:
    get:
      consumes:
      - application/json
      description: 根据设备ID获取设备的详细信息
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.ErrorResponse'
      summary: 获取设备详情
      tags:
      - 设备管理
...
  /fe-v1/elastic-scaling/orders/{id}/devices:
    get:
      consumes:
      - application/json
      description: 获取指定订单关联的设备列表
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.ErrorResponse'
      summary: 获取订单关联的设备
      tags:
      - 弹性伸缩订单
...
  /resource-pool/matching-policies:
    get:
      consumes:
      - application/json
      description: 获取资源池设备匹配策略列表
      parameters:
      - description: 页码，默认1
        in: query
        name: page
        type: integer
      - description: 每页数量，默认10
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.Response'
      summary: 获取资源池设备匹配策略列表
      tags:
      - 资源池设备匹配策略
    post:
      consumes:
      - application/json
      description: 创建资源池设备匹配策略
      parameters:
      - description: 创建策略请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateResourcePoolDeviceMatchingPolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.Response'
      summary: 创建资源池设备匹配策略
      tags:
      - 资源池设备匹配策略
  /resource-pool/matching-policies/{id}:
    delete:
      consumes:
      - application/json
      description: 删除资源池设备匹配策略
      parameters:
      - description: 策略ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.Response'
      summary: 删除资源池设备匹配策略
      tags:
      - 资源池设备匹配策略
    get:
      consumes:
      - application/json
      description: 获取资源池设备匹配策略详情
      parameters:
      - description: 策略ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.Response'
      summary: 获取资源池设备匹配策略详情
      tags:
      - 资源池设备匹配策略
    put:
      consumes:
      - application/json
      description: 更新资源池设备匹配策略
      parameters:
      - description: 策略ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新策略请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateResourcePoolDeviceMatchingPolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.Response'
      summary: 更新资源池设备匹配策略
      tags:
      - 资源池设备匹配策略
  /resource-pool/matching-policies/{id}/status:
    put:
      consumes:
      - application/json
      description: 更新资源池设备匹配策略状态
      parameters:
      - description: 策略ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新状态请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateResourcePoolDeviceMatchingPolicyStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.Response'
      summary: 更新资源池设备匹配策略状态
      tags:
      - 资源池设备匹配策略
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"